This does a few cleanups and fixes a hang in MPI on intel.

+++ libgadget/domain.h
@@ -12,7 +12,7 @@ struct topnode_data
 {
     peano_t StartKey;		/*!< first Peano-Hilbert key in top-level node */
     int Daughter;			/*!< index of first daughter cell (out of 8) of top-level node */
-    short int Shift;    /*!< level = log 2 of number of Peano-Hilbert mesh-cells represented by top-level node */
+    int Shift;    /*!< level = log 2 of number of Peano-Hilbert mesh-cells represented by top-level node */
     int Leaf ;			/*!< if the node is a leaf, this gives its index in topleaf_data */
 };
